De-orphan any instances which are in the same package
